refactor: Move typedefs to solo12_types.hpp
Move the robot_interfaces typedefs to a separate header file
`solo12_types.hpp`.  This will make it easier to add alternative drivers
in separate files.

Also rename the existing `types.hpp` to `basic_types.hpp`.
