+++ include/robot_interfaces_solo/solo12_action.hpp
@@ -14,7 +14,7 @@
 
 #include <robot_interfaces/loggable.hpp>
 
-#include "types.hpp"
+#include "basic_types.hpp"
 
 namespace robot_interfaces_solo
 {
